+++ lib/devise/multi_email/models/authenticatable.rb
@@ -42,7 +42,7 @@ module Devise
 
         # Sets the default email address of the user.
         def email=(new_email)
-          multi_email.change_primary_email_to(new_email, force_primary: true)
+          multi_email.change_primary_email_to(new_email, allow_unconfirmed: true)
         end
       end