+++ lib/devise/multi_email/parent_model_manager.rb
@@ -15,7 +15,7 @@ module Devise
       def login_email_record
         if @parent_record.current_login_email.present?
           formatted_email = format_email(@parent_record.current_login_email)
-          filtered_emails.find{ |item| item.email == formatted_email }
+          filtered_emails.find { |item| item.email == formatted_email }
         end
       end
 
@@ -25,19 +25,19 @@ module Devise
       end
       alias_method Devise::MultiEmail.primary_email_method_name, :primary_email_record
 
-      # :force_primary option sets this email record to primary
+      # :allow_unconfirmed option sets this email record to primary
       # :skip_confirmations option confirms this email record (without saving)
       # @see `set_primary_record_to`
       def change_primary_email_to(new_email, options = {})
         # mark none as primary when set to nil
         if new_email.nil?
-          filtered_emails.each{ |item| item.primary = false }
+          filtered_emails.each { |item| item.primary = false }
 
         # select or build an email record
         else
           record = find_or_build_for_email(new_email)
 
-          if record.try(:confirmed?) || primary_email_record.nil? || options[:force_primary]
+          if record.try(:confirmed?) || primary_email_record.nil? || options[:allow_unconfirmed]
             set_primary_record_to(record, options)
           end
         end
@@ -52,7 +52,7 @@ module Devise
 
       def find_or_build_for_email(email)
         formatted_email = format_email(email)
-        record = filtered_emails.find{ |item| item.email == formatted_email }
+        record = filtered_emails.find { |item| item.email == formatted_email }
         record || emails.build(email: formatted_email)
       end
 
@@ -66,11 +66,11 @@ module Devise
       end
 
       def confirmed_emails
-        filtered_emails.select{ |record| record.try(:confirmed?) }
+        filtered_emails.select { |record| record.try(:confirmed?) }
       end
 
       def unconfirmed_emails
-        filtered_emails.reject{ |record| record.try(:confirmed?) }
+        filtered_emails.reject { |record| record.try(:confirmed?) }
       end
 
     protected
@@ -78,7 +78,7 @@ module Devise
       # :skip_confirmations option confirms this email record (without saving)
       def set_primary_record_to(record, options = {})
         # Toggle primary flag for all emails
-        filtered_emails.each{ |other| other.primary = (other.email == record.email) }
+        filtered_emails.each { |other| other.primary = (other.email == record.email) }
 
         if options[:skip_confirmations]
           record.try(:skip_confirmation!)