+++ spec/rails_app/config/application.rb
@@ -10,7 +10,7 @@ require 'devise/multi_email'
 module RailsApp
   class Application < Rails::Application
     # Add additional load paths for your own custom dirs
-    config.autoload_paths.reject!{ |p| p =~ /\/app\/(\w+)$/ && !%w(controllers helpers mailers models views).include?($1) }
+    config.autoload_paths.reject! { |p| p =~ /\/app\/(\w+)$/ && !%w(controllers helpers mailers models views).include?($1) }
 
     # Configure generators values. Many other options are available, be sure to check the documentation.
     # config.generators do |g|