+++ spec/rails_app/config/initializers/devise-multi_email.rb
@@ -1,6 +1,6 @@
 
 Devise::MultiEmail.configure do |config|
-  #config.configure_autosave = false
+  #config.autosave_emails = false
   #config.primary_email_method_name = :primary_email
   #config.only_login_with_primary_email = false
 end