+++ spec/support/features.rb
@@ -16,7 +16,7 @@ module RailsTestHelpers
     email_address = user.multi_email.format_email(options[:email] || "user_#{SecureRandom.hex}@test.com")
     user.email = email_address
 
-    email = user.emails.to_a.find{ |record| record.email == email_address}
+    email = user.emails.to_a.find { |record| record.email == email_address }
     email.update_attribute(:confirmation_sent_at, options[:confirmation_sent_at]) if options[:confirmation_sent_at]
 
     if options[:confirm] == false