+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 * Refactored to simplify some logic and start moving toward mimicking default Devise lifecycle behavior
 * Added `Devise::MultiEmail.only_login_with_primary_email` option to restrict login to only primary emails
-* Added `Devise::MultiEmail.configure_autosave` option to automatically enable `autosave` on "emails" association
+* Added `Devise::MultiEmail.autosave_emails` option to automatically enable `autosave` on "emails" association
 
 ### 2.0.0 - 2017-05-12